complete search function, chang type of price to INT, type of id to long for consistent

+++ app/src/main/java/com/example/projectdemo04/model/Book.java
@@ -154,7 +154,7 @@ public class Book implements Serializable {
         this.rating = rating;
     }
 
-    public double getPrice() {
+    public int getPrice() {
         return price;
     }
